Clean compile on windows
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1948 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/wp64/kernel/parser/InstructionParser.cpp
@@ -446,7 +446,7 @@ RexxInstruction *RexxSource::createDoLoop(RexxInstructionDo *newDo, bool isLoop)
 /* Function:  Create a new DO translator object                               */
 /******************************************************************************/
 {
-    int _position = markPosition();     // get a reset position before getting next token
+    size_t _position = markPosition();   // get a reset position before getting next token
     // We've already figured out this is either a LOOP or a DO, now we need to
     // decode what else is going on.
     RexxToken *token = nextReal();
@@ -969,7 +969,7 @@ RexxInstruction *RexxSource::guardNew()
   RexxToken  *token;                   /* current working token             */
   bool        on_off = false;          /* ON or OFF version                 */
   RexxObject *_expression;             /* WHEN expression                   */
-  int         variable_count;          /* count of variables                */
+  size_t      variable_count;          /* count of variables                */
 
   _expression = OREF_NULL;             /* default no expression             */
   variable_list = OREF_NULL;           /* no variable either                */
@@ -1684,7 +1684,7 @@ RexxInstruction *RexxSource::raiseNew()
   RexxObject            *_expression;  /* condition extra expression        */
   RexxObject            *description;  /* description expression            */
   RexxObject            *additional;   /* additional expression             */
-  int                    arrayCount;   /* array expressions                 */
+  size_t                 arrayCount;   /* array expressions                 */
   RexxObject            *result;       /* result expression                 */
   int                    _keyword;     /* type of condition found           */
   RexxQueue             *saveQueue;    /* temporary save queue              */
@@ -2125,8 +2125,8 @@ RexxInstruction *RexxSource::traceNew()
   RexxString *value;                   /* string value of symbol            */
   RexxObject *_expression;             /* trace expression                  */
   RexxObject *newObject;               /* newly created object              */
-  int         setting;                 /* new trace setting                 */
-  int         debug;                   /* new debug setting                 */
+  size_t      setting;                 /* new trace setting                 */
+  size_t      debug;                   /* new debug setting                 */
   wholenumber_t debug_skip;            /* amount to skip                    */
   size_t      debug_flags;             /* current debug flags               */
